refactor: Rename utils files to match naming convention
