+++ src/content/utils/__tests__/is-object.spec.ts
@@ -1,4 +1,4 @@
-import { isObject } from '../isObject';
+import { isObject } from '../is-object';
 
 describe('isObject', () => {
   it('returns false when value is undefined', () => {
