+++ src/content/utils/build-collection-full-name.ts
