+++ src/content/utils/create-xul-element.ts
