+++ src/content/utils/get-localized-string.ts
