+++ src/content/utils/get-xul-element-by-id.ts
