+++ src/content/utils/has-error-stack.ts
