+++ src/content/utils/index.ts
@@ -1,12 +1,12 @@
-export { buildCollectionFullName } from './buildCollectionFullName';
-export { chunkString } from './chunkString';
-export { createHTMLElement } from './createHTMLElement';
-export { createXULElement } from './createXULElement';
-export { getAllCollectionItems } from './getAllCollectionItems';
-export { getDOMParser } from './getDOMParser';
-export { getLocalizedString } from './getLocalizedString';
-export { getXULElementById } from './getXULElementById';
-export { hasErrorStack } from './hasErrorStack';
-export { isObject } from './isObject';
-export { keyValue } from './keyValue';
+export { buildCollectionFullName } from './build-collection-full-name';
+export { chunkString } from './chunk-string';
+export { createHTMLElement } from './create-html-element';
+export { createXULElement } from './create-xul-element';
+export { getAllCollectionItems } from './get-all-collection-items';
+export { getDOMParser } from './get-dom-parser';
+export { getLocalizedString } from './get-localized-string';
+export { getXULElementById } from './get-xul-element-by-id';
+export { hasErrorStack } from './has-error-stack';
+export { isObject } from './is-object';
+export { keyValue } from './key-value';
 export { log } from './log';
