+++ src/content/utils/is-object.ts
